@@ -32,37 +38,45 @@
     "chai": "^4.0.2",
     "chalk": "^2.1.0",
     "css-loader": "^0.28.1",
-    "enzyme": "^3.0.0",
-    "eslint": "^4.1.0",
-    "eslint-config-kreativgebiet": "^0.2.2",
-    "eslint-plugin-flowtype": "^2.34.0",
+    "enzyme": "^3.3.0",
+    "enzyme-adapter-react-16": "^1.1.1",
+    "eslint": "^4.19.1",
+    "eslint-config-airbnb": "16.1.0",
+    "eslint-config-prettier": "^2.9.0",
+    "eslint-import-resolver-babel-module": "^4.0.0",
+    "eslint-plugin-flowtype": "^2.46.3",
+    "eslint-plugin-import": "^2.7.0",
+    "eslint-plugin-jsx-a11y": "^6.0.2",
+    "eslint-plugin-prettier": "^2.6.0",
+    "eslint-plugin-react": "^7.4.0",
     "file-loader": "^0.11.1",
-    "flow-bin": "^0.51.1",
-    "html-webpack-plugin": "^2.30.1",
-    "img-loader": "^2.0.0",
+    "flow-bin": "^0.69.0",
+    "flow-coverage-report": "^0.5.0",
+    "html-webpack-plugin": "^3.2.0",
+    "img-loader": "^3.0.0",
     "ip": "^1.1.5",
+    "jest": "^22.4.3",
     "jsdom": "^11.0.0",
+    "madge": "^3.0.1",
     "minimist": "^1.2.0",
-    "mocha": "^3.4.2",
-    "mocha-webpack": "^0.7.0",
     "null-loader": "^0.1.1",
-    "react-hot-loader": "next",
+    "prettier": "^1.12.1",
     "react-test-renderer": "^16.0.0",
-    "sinon": "^3.0.0",
     "style-loader": "0.19.0",
     "url-loader": "^1.0.0",
-    "webpack": "^3.0.0",
-    "webpack-dev-server": "^2.7.1",
-    "webpack-hot-middleware": "^2.7.1",
+    "webpack": "^4.8.3",
+    "webpack-cli": "^2.1.3",
+    "webpack-dev-middleware": "^3.1.3",
+    "webpack-dev-server": "^3.1.4",
+    "webpack-hot-middleware": "^2.22.2",
     "webpack-node-externals": "^1.6.0"
   },
   "dependencies": {
-    "aphrodite": "^1.2.4",
-    "autobind-decorator": "^2.1.0",
+    "aphrodite": "^2.2.0",
     "classnames": "^2.2.3",
+    "coveralls": "^3.0.1",
     "express": "^4.15.3",
     "history": "^4.6.1",
-    "immutable": "^3.8.1",
     "invariant": "^2.2.2",
     "lodash": "^4.17.4",
     "normalizr": "^3.2.2",
