@@ -1,68 +1,50 @@
 // @flow
+import React from 'react'
+import { Link } from 'react-router-dom'
+import { css } from 'aphrodite'
 
-import React from 'react';
-import PropTypes from 'prop-types';
-import { Link } from 'react-router-dom';
-import { css } from 'aphrodite';
-import has from 'lodash/has';
-import invariant from 'invariant';
+import { spinConfig as defaultSpinConfig } from 'app/config.js'
+import Spin from 'components/Spin'
 
-import { spinConfig as defaultSpinConfig } from '@root/config.js';
-import Spin from '@components/Spin';
-import styles from './style.js';
+import styles from './style.js'
+import type { Props } from './types.js'
 
-export default function Button(allProps: Object) {
-  let Tag = 'button';
-  const { loading, spinConfig, ...props } = allProps;
+export default function Button(allProps: Props): React$Node {
+  let Tag = 'button'
+  const { loading, spinConfig, ...props } = allProps
 
-  if (has(props, 'to')) {
-    Tag = Link;
+  if (Object.prototype.hasOwnProperty.call(props, 'to')) {
+    Tag = Link
 
-    delete props.href;
-    delete props.onClick;
-  } else if (has(props, 'href')) {
-    Tag = 'a';
+    delete props.href
+    delete props.onClick
+  } else if (Object.prototype.hasOwnProperty.call(props, 'href')) {
+    Tag = 'a'
 
-    delete props.to;
-    delete props.onClick;
+    delete props.to
+    delete props.onClick
   }
 
   return (
     <Tag
       disabled={loading}
-      className={css(styles.main, loading && styles.main__loading)}
+      className={css(styles.button, !!loading && styles.button__loading)}
       {...props}
     >
       {/* Implement the spinner for loading activity */}
-      <span className={css(styles.main__spinner, loading && styles.main__loading__active)}>
+      <span className={css(styles.button__spinner, !!loading && styles['button__spinner--active'])}>
         <Spin stopped={!loading} config={{ ...defaultSpinConfig, ...spinConfig }} />
       </span>
 
       {/* Implement the children (content) of the button */}
-      <span className={css(styles.main__content, loading && styles.main__loading__active)}>
+      <span className={css(styles.button__content, !!loading && styles['button__spinner--active'])}>
         {props.children}
       </span>
     </Tag>
-  );
+  )
 }
 
 Button.defaultProps = {
   spinConfig: defaultSpinConfig,
   loading: false,
-  to: null,
-  href: null,
-  onClick: () =>
-    invariant(false, 'No `onClick`, `to` or `href` was defined. Used fallback, with a noop.'),
-};
-
-Button.propTypes = {
-  to: PropTypes.string,
-  loading: PropTypes.bool,
-  href: PropTypes.string,
-  onClick: PropTypes.func,
-  spinConfig: PropTypes.object,
-  children: PropTypes.oneOfType([
-    PropTypes.arrayOf(PropTypes.node),
-    PropTypes.node,
-  ]).isRequired,
-};
+}
