@@ -1,3 +1,4 @@
+// @flow
 
-export Header from './Header';
-export Button from './Button';
+export { default as Header } from './Header'
+export { default as Button } from './Button'
