@@ -1,4 +1,7 @@
+// @flow
+import createHistory from 'history/createBrowserHistory'
+import type { BrowserHistory } from 'history/createBrowserHistory'
 
-import createHistory from 'history/createBrowserHistory';
+const history: BrowserHistory = createHistory()
 
-export default createHistory();
+export default history
