@@ -1,10 +1,12 @@
+// @flow
+import { combineReducers } from 'redux'
+import { routerReducer } from 'react-router-redux/reducer'
+import type { CombinedReducer } from 'redux'
 
-import { combineReducers } from 'redux';
-import { routerReducer } from 'react-router-redux';
+import { reducer as tasksReducer } from './tasks'
+import type { Action } from './types'
 
-import { tasksReducer } from './tasks';
-
-export default combineReducers({
+export const rootReducer: CombinedReducer<any, Action> = combineReducers({
   router: routerReducer,
   tasks: tasksReducer,
-});
+})
