@@ -1,3 +1,3 @@
-
-export const CREATE_TASK = 'CREATE_TASK';
-export const REMOVE_TASK = 'REMOVE_TASK';
+// @flow
+export const CREATE_TASK: string = 'CREATE_TASK'
+export const REMOVE_TASK: string = 'REMOVE_TASK'
