@@ -1,13 +1,8 @@
+// @flow
+import * as actions from './actions'
+import * as actionTypes from './actionTypes'
 
-import * as tasksActions from './actions';
-
-export { tasksActions };
-export {
-  CREATE_TASK,
-  REMOVE_TASK,
-} from './actionTypes';
-export { tasksReducer } from './reducer';
-export { Task } from './task';
-export {
-  getTasks,
-} from './selectors';
+export { actions }
+export { actionTypes }
+export { reducer } from './reducer'
+export { getTasks } from './selectors'
