@@ -1,17 +1,15 @@
-/* eslint-disable no-case-declarations */
-import { Record, List } from 'immutable';
-
-import { CREATE_TASK, REMOVE_TASK } from './actionTypes';
-import { Task } from './task';
+// @flow
+import { CREATE_TASK, REMOVE_TASK } from './actionTypes'
+import type { State, Action } from './types'
 
 /*
  * Immutable Record for the initial state tree for `tasks`. Beside being
  * immutable it gives us some cool features like merging and stuff.
  */
-export const TasksState = new Record({
+export const initialState: State = {
   loading: true,
-  list: new List(),
-});
+  list: [],
+}
 
 /*
  * The reducer is called *every* time redux receives a new action.
@@ -19,16 +17,13 @@ export const TasksState = new Record({
  *  - CREATE_TASK
  *  - REMOVE_TASK
  */
-export function tasksReducer(state = new TasksState(), { type, ...action }) {
+export const reducer = (state: State = initialState, { type, ...action }: Action): State => {
   switch (type) {
     case CREATE_TASK:
-      const newTask = new Task(action);
-      return state.update('list', list => list.push(newTask));
+      return { ...state, list: action.payload ? [...state.list, action.payload] : state.list }
     case REMOVE_TASK:
-      return state.update('list', list =>
-        list.filter(task => task.get('id') === action.id),
-      );
+      return { ...state, list: state.list.filter(task => task.id !== action.id) }
     default:
-      return state;
+      return state
   }
 }
