@@ -1,33 +1,16 @@
+import path from 'path'
+import paths from './paths'
 
-import path from 'path';
-import paths from './paths';
-
-const { NODE_ENV } = process.env;
+const { NODE_ENV } = process.env
+const isDev = NODE_ENV === 'development'
 
 module.exports = {
   context: paths.appBuild,
-  entry: {
-    jsx: [
-      NODE_ENV === 'development' && 'react-hot-loader/patch',
-      './index.js',
-    ].filter(entry => !!entry),
-    vendor: [
-      'react',
-      'react-dom',
-      'react-redux',
-      'react-router',
-      'react-router-dom',
-      'react-router-redux',
-      'redux',
-      'history',
-      'immutable',
-      'classnames',
-      'normalizr',
-    ],
-  },
+  entry: [isDev && 'react-hot-loader/patch', './index.js'].filter(e => e),
   output: {
     path: paths.outputPath,
     filename: paths.outputName,
+    publicPath: '/',
   },
   module: {
     strictExportPresence: NODE_ENV === 'production',
@@ -50,15 +33,12 @@ module.exports = {
   },
   resolve: {
     alias: {
-      '@root': path.resolve(paths.appBuild),
-      '@assets': path.resolve(paths.appBuild, 'assets'),
-      '@pages': path.resolve(paths.appBuild, 'pages'),
-      '@components': path.resolve(paths.appBuild, 'components'),
-      '@modules': path.resolve(paths.appBuild, 'modules'),
+      app: path.resolve(paths.appBuild),
+      assets: path.resolve(paths.appBuild, 'assets'),
+      pages: path.resolve(paths.appBuild, 'pages'),
+      components: path.resolve(paths.appBuild, 'components'),
+      modules: path.resolve(paths.appBuild, 'modules'),
     },
-    modules: [
-      path.resolve('./'),
-      'node_modules',
-    ],
+    modules: [path.resolve('./'), 'node_modules'],
   },
-};
+}
