@@ -1,7 +1,6 @@
-
-import webpack from 'webpack';
-import HtmlWebpackPlugin from 'html-webpack-plugin';
-import baseConfig from './webpack.config.common.js';
+import webpack from 'webpack'
+import HtmlWebpackPlugin from 'html-webpack-plugin'
+import baseConfig from './webpack.config.common.js'
 
 module.exports = {
   ...baseConfig,
@@ -15,7 +14,6 @@ module.exports = {
   },
   plugins: [
     new webpack.NamedModulesPlugin(),
-    new webpack.optimize.CommonsChunkPlugin({ name: 'vendor', filename: 'vendor.bundle.js' }),
     new webpack.DefinePlugin({
       'process.env': {
         NODE_ENV: JSON.stringify('development'),
@@ -27,4 +25,4 @@ module.exports = {
       template: 'index.html',
     }),
   ],
-};
+}
