@@ -1,13 +1,13 @@
 /* eslint no-console: 0 */
 
-const express = require('express');
+const express = require('express')
 
-const app = express();
+const app = express()
 
 app.get('/', (req, res) => {
-  res.send('Hello World!');
-});
+  res.send('Hello World!')
+})
 
 app.listen(3000, () => {
-  console.log('Example app listening on port 3000!');
-});
+  console.log('Example app listening on port 3000!')
+})
